requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
scikit-learn==0.24.2
mlflow==1.20.2
joblib==1.0.1
pydantic==1.8.2
lifetime
numpy